Create Out-AllMembersGrid
Probably I should not pipe into Out-GridView in line 69, but it was my use-case at the time I built this. This cmdlet normalizes object headers in collections. Normally in PowerShell, the first object in a collection sets the headers for all subsequent objects in the collections even if they have different headers. This cmdlet gathers all the headers in the collection and attaches them to each object, so that they are displayed correctly.