Create Export-PowerShellDataFile
The opposite of Import-PowerShellDataFile. This file contains 2 functions:

The first function is ConvertTo-PowerShellDataFile. This converts any object into PSDataFile format, e.g., `cat file.json | ConvertFrom-Json | ConvertTo-PowerShellDataFile | Set-Content -Path new.psd1`

The second function is the Export-PowerShellDataFile, which is essentially a wrapper for `ConvertTo-PowerShellDataFile | Set-Content -Path <outfile>`.

The main workhorse is obviously ConvertTo-PowerShellDataFile, but you can't do anything with the PSD1 format except load it back again with Invoke-Expression. So most likely you are simply interested in writing it to a file so you can reload it later. This is where the Export wrapper comes in, to take your objects and directly serialize them in the native PS config format.

The ConvertTo-PowerShellDataFile function is inspired by Dave Wyatt (assuming they are the original author) at https://stackoverflow.com/a/34383413/6076137