Update Import-Yaml
Added a small function, ConvertFrom-Yaml, similar to ConvertFrom-Json, just for the sake of completeness and to increase the versatility of this file. However, the function is quick and dirty and serializes to a temporary file. It would be simple to update the yaml class to not require a file, but then its name FileYaml wouldn't make much sense if it doesn't need a file. It's late. I'll come back to fix this triviality when I have a motive to.