Slimmed down the Context mapping sync function workload
It will now only map GKE relevant project IDs
This saves considerable time.

Also tried a small refactoring to hopefully fix
the stability, as sometimes it hangs at the start.
I think there was a variable that was not always
being resolved properly, but will keep an eye on it.