updated modules: kubectlutils and gcloudutils for faster loading
kubectlutils initializes with only 1 kubectl call.
2-3 others were removed. Currently the full api
resources are no longer cached, but they weren't
being used yet so fixing it is a To-Do for later.

gcloudutils no longer makes a gcloud call to get
the current active project. Instead, PowerShell
serializes the location of the configuration file
directory and makes its own lookup of the active
project based on the last-edited config file. This
is much faster than a gcloud call which takes up to
a second.

For now, gcloud tab completion no longer automatically
initializes. Once the module is ready, I'll reactivate
it again, or maybe set an event that when gcloud is
called, it imports the completion.

Overall, the initialization time of the shell has been
reduced from 10-15 seconds down to 3-5 seconds (or at
least it looks this way at the moment).

cliutils also has updated its git wrapper.